Change package name back to edu.rpi.rpimobile
Prepare for release v1.0 by changing package name
Add Map fragment into preferences' startup screen pane
Fix naming convention of TweetObject class and all references
  so that it is in camelcase
