+++ RPIMobile/bin/jarlist.cache
@@ -1,0 +1,5 @@
+# cache for current jar dependency. DO NOT EDIT.
+# format is <lastModified> <length> <SHA-1> <path>
+# Encoding is UTF-8
+1385876189502 621451 5896b0a4e377ac4242eb2bc785220c1c4fc052f4 C:\Users\piechp\Documents\GitHub\RPIMobile-Android\RPIMobile\libs\android-support-v4.jar
+1385876189502 621451 5896b0a4e377ac4242eb2bc785220c1c4fc052f4 C:\Users\piechp\Documents\GitHub\RPIMobile-Android\actionbarsherlock\libs\android-support-v4.jar
