+++ RPIMobile/src/edu/rpi/rpimobile/MapFragment.java
@@ -8,7 +8,7 @@
  *              a Google Map view.
  */
 
-package rpi.edu.rpimobile;
+package edu.rpi.rpimobile;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -16,7 +16,7 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.ArrayList;
 
-import rpi.edu.rpimobile.model.MapLocation;
+import edu.rpi.rpimobile.model.MapLocation;
 
 import android.database.Cursor;
 import android.database.SQLException;
