+++ RPIMobile/src/edu/rpi/rpimobile/model/Building.java
@@ -1,4 +1,4 @@
-package rpi.edu.rpimobile.model;
+package edu.rpi.rpimobile.model;
 
 //Class for storing all of the variables associated with a laundry call
 public class Building{
